+++ custom_components/ai_thermostat/climate.py
@@ -18,11 +18,6 @@ from homeassistant.components.climate.const import (
 	HVAC_MODE_HEAT,
 	HVAC_MODE_OFF,
 	SUPPORT_TARGET_TEMPERATURE,
-	SERVICE_SET_TEMPERATURE,
-	SERVICE_SET_HVAC_MODE,
-)
-from homeassistant.components.number.const import (
-	SERVICE_SET_VALUE,
 )
 from homeassistant.components.recorder import history
 from homeassistant.const import (
@@ -41,7 +36,7 @@ from homeassistant.helpers.event import (
 from homeassistant.helpers.reload import async_setup_reload_service
 from homeassistant.helpers.restore_state import RestoreEntity
 
-from custom_components.ai_thermostat.helpers import check_float, convert_decimal
+from custom_components.ai_thermostat.helpers import check_float, convert_decimal, set_trv_values
 from custom_components.ai_thermostat.models.models import convert_inbound_states, convert_outbound_states
 from . import DOMAIN, PLATFORMS
 
@@ -612,22 +607,22 @@ class AIThermostat(ClimateEntity, RestoreEntity, ABC):
 		"""Maintenance of the TRV valve."""
 		self.ignore_states = True
 		if self.hass.states.get(self.heater_entity_id).attributes.get('valve_position'):
-			await self.set_trv_values('valve_position', 255)
+			await set_trv_values(self,'valve_position', 255)
 			await asyncio.sleep(60)
-			await self.set_trv_values('valve_position', 0)
+			await set_trv_values(self,'valve_position', 0)
 			await asyncio.sleep(60)
-			await self.set_trv_values('valve_position', 255)
+			await set_trv_values(self,'valve_position', 255)
 			await asyncio.sleep(60)
-			await self.set_trv_values('valve_position', 0)
+			await set_trv_values(self,'valve_position', 0)
 			await asyncio.sleep(60)
 		else:
-			await self.set_trv_values('temperature', 30)
+			await set_trv_values(self,'temperature', 30)
 			await asyncio.sleep(60)
-			await self.set_trv_values('temperature', 5)
+			await set_trv_values(self,'temperature', 5)
 			await asyncio.sleep(60)
-			await self.set_trv_values('temperature', 30)
+			await set_trv_values(self,'temperature', 30)
 			await asyncio.sleep(60)
-			await self.set_trv_values('temperature', 5)
+			await set_trv_values(self,'temperature', 5)
 			await asyncio.sleep(60)
 		self.ignore_states = False
 		await self._async_control_heating()
@@ -727,21 +722,21 @@ class AIThermostat(ClimateEntity, RestoreEntity, ABC):
 							self.hass.states.get(self.heater_entity_id).attributes.get('device').get('friendlyName')
 					)
 					
-					# FIXME: this if is waaaay to complex
+					# Using on temperature based calibration, dont update the temp if its the same
 					if self.calibration_type == 1 and float(self.hass.states.get(self.heater_entity_id).attributes.get('current_heating_setpoint')) != float(calibration):
-						await self.set_trv_values('temperature', float(calibration))
+						await set_trv_values(self,'temperature', float(calibration))
 					
-					# FIXME: this if is waaaay to complex
+					# Using on local calbiration, dont update the temp if its off, some TRV changed to 5°C when off after a while, don't update the temp
 					if self.calibration_type == 0 and not self.window_open and converted_hvac_mode != HVAC_MODE_OFF and float(current_heating_setpoint) != 5.0 and self.call_for_heat:
-						await self.set_trv_values('temperature', float(current_heating_setpoint))
+						await set_trv_values(self,'temperature', float(current_heating_setpoint))
 					
-					# Calibration stuff
+					# Using on local calbiration, update only if the TRV is not in window open mode
 					if self.calibration_type == 0 and not self.window_open and do_calibration:
-						await self.set_trv_values('local_temperature_calibration', calibration)
+						await set_trv_values(self,'local_temperature_calibration', calibration)
 					
-					# FIXME: this if is waaaay to complex
-					if has_real_mode and (converted_hvac_mode != self.hass.states.get(self.heater_entity_id).attributes.get('system_mode') or converted_hvac_mode == HVAC_MODE_OFF):
-						await self.set_trv_values('system_mode', converted_hvac_mode)
+					# Only set the system mode if the TRV has this option
+					if has_real_mode:
+						await set_trv_values(self,'system_mode', converted_hvac_mode)
 					
 					self.ignore_states = False
 					
@@ -757,25 +752,6 @@ class AIThermostat(ClimateEntity, RestoreEntity, ABC):
 					_LOGGER.debug("ai_thermostat entity not ready or device is currently not supported")
 					_LOGGER.debug("fatal %s", fatal)
 					self.ignore_states = False
-	
-	async def set_trv_values(self, key, value):
-		if key == 'temperature':
-			await self.hass.services.async_call('climate', SERVICE_SET_TEMPERATURE, {'entity_id': self.heater_entity_id, 'temperature': value}, blocking=True)
-		elif key == 'system_mode':
-			await self.hass.services.async_call('climate', SERVICE_SET_HVAC_MODE, {'entity_id': self.heater_entity_id, 'hvac_mode': value}, blocking=True)
-		elif key == 'local_temperature_calibration':
-			local_calibration_entity = self.heater_entity_id.replace('climate.', 'number.') + '_local_temperature_calibration'
-			max_calibration = self.hass.states.get(local_calibration_entity).attributes.get('max')
-			min_calibration = self.hass.states.get(local_calibration_entity).attributes.get('min')
-			if value > max_calibration:
-				value = max_calibration
-			if value < min_calibration:
-				value = min_calibration
-			await self.hass.services.async_call('number', SERVICE_SET_VALUE, {'entity_id': local_calibration_entity, 'value': value}, blocking=True)
-		elif key == 'valve_position':
-			valve_position_entity = self.heater_entity_id.replace('climate.', 'number.') + '_valve_position'
-			await self.hass.services.async_call('number', SERVICE_SET_VALUE, {'entity_id': valve_position_entity, 'value': value}, blocking=True)
-		await asyncio.sleep(1)
 	
 	def check_weather_prediction(self):
 		"""
